Separate layer repository and service
Fixes #158

Separate the repository and service layers in the horext/app-api project.

* **Repository Layer**: Add repository interfaces and implementations for `ClassSession`, `HourlyLoad`, `OrganizationUnit`, `Schedule`, `ScheduleSubject`, `StudyPlan`, and `Subject` in the `src/main/kotlin/io/octatec/horext/api/repository` directory.
* **Service Layer**: Modify service implementations to use the newly created repository classes for data access. Inject the appropriate repository into each service implementation and delegate data access methods to the repository.
