+++ src/main/kotlin/io/octatec/horext/api/repository/ClassSessionRepository.kt
@@ -1,0 +1,8 @@
+package io.octatec.horext.api.repository
+
+import io.octatec.horext.api.domain.ClassSession
+
+interface ClassSessionRepository {
+    fun findByScheduleId(scheduleId: Long): List<ClassSession>
+    fun findByScheduleIds(scheduleIds: List<Long>): List<ClassSession>
+}
