+++ src/main/kotlin/io/octatec/horext/api/repository/HourlyLoadRepository.kt
@@ -1,0 +1,7 @@
+package io.octatec.horext.api.repository
+
+import io.octatec.horext.api.domain.HourlyLoad
+
+interface HourlyLoadRepository {
+    fun getLatestByFaculty(facultyId: Long): HourlyLoad
+}
