+++ src/main/kotlin/io/octatec/horext/api/repository/OrganizationUnitRepository.kt
@@ -1,0 +1,9 @@
+package io.octatec.horext.api.repository
+
+import io.octatec.horext.api.domain.OrganizationUnit
+
+interface OrganizationUnitRepository {
+    fun getAllSpeciality(): List<OrganizationUnit>
+    fun getAllFaculty(): List<OrganizationUnit>
+    fun getAllSpecialityByFacultyId(id: Long): List<OrganizationUnit>
+}
