+++ src/main/kotlin/io/octatec/horext/api/repository/ScheduleRepository.kt
@@ -1,0 +1,7 @@
+package io.octatec.horext.api.repository
+
+import io.octatec.horext.api.domain.Schedule
+
+interface ScheduleRepository {
+    fun findBySubjectIdAndHourlyLoadId(subjectId: Long, hourlyLoadId: Long): List<Schedule>
+}
