+++ src/main/kotlin/io/octatec/horext/api/repository/StudyPlanRepositoryImpl.kt
@@ -1,0 +1,24 @@
+package io.octatec.horext.api.repository
+
+import io.octatec.horext.api.domain.StudyPlan
+import io.octatec.horext.api.domain.StudyPlans
+import io.octatec.horext.api.domain.OrganizationUnits
+import org.jetbrains.exposed.sql.and
+import org.jetbrains.exposed.sql.selectAll
+import org.springframework.stereotype.Repository
+
+@Repository
+class StudyPlanRepositoryImpl : StudyPlanRepository {
+
+    override fun getAllStudyPlan(): List<StudyPlan> {
+        return StudyPlans.selectAll().map { row -> StudyPlans.createEntity(row) }
+    }
+
+    override fun getStudyPlanById(id: Long): List<StudyPlan> {
+        val sp = StudyPlans
+        val ou = OrganizationUnits
+        return sp.leftJoin(ou)
+        .select( sp.columns + ou.columns)
+        .where { (sp.id eq id) }.map { row -> sp.createEntity(row) }
+    }
+}
