+++ src/main/kotlin/io/octatec/horext/api/service/HourlyLoadServiceImpl.kt
@@ -1,28 +1,15 @@
 package io.octatec.horext.api.service
 
-import io.octatec.horext.api.domain.AcademicPeriodOrganizationUnits
 import io.octatec.horext.api.domain.HourlyLoad
-import io.octatec.horext.api.domain.HourlyLoads
-import org.jetbrains.exposed.sql.and
+import io.octatec.horext.api.repository.HourlyLoadRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.time.Instant
 
 @Service
 @Transactional
-class HourlyLoadServiceImpl() : HourlyLoadService {
+class HourlyLoadServiceImpl(private val hourlyLoadRepository: HourlyLoadRepository) : HourlyLoadService {
 
     override fun getLatestByFaculty(facultyId: Long): HourlyLoad {
-        val hl = HourlyLoads
-        val apou = AcademicPeriodOrganizationUnits
-        return hl
-            .innerJoin(apou)
-            .select(hl.columns + apou.columns)
-            .where {
-                (apou.organizationUnitId eq facultyId) and
-                        (hl.publishedAt lessEq Instant.now())
-            }
-            .map { row -> hl.createEntity(row) }.first()
-
+        return hourlyLoadRepository.getLatestByFaculty(facultyId)
     }
 }
