+++ src/main/kotlin/io/octatec/horext/api/service/StudyPlanServiceImpl.kt
@@ -1,26 +1,19 @@
 package io.octatec.horext.api.service
 
 import io.octatec.horext.api.domain.StudyPlan
-import io.octatec.horext.api.domain.StudyPlans
-import io.octatec.horext.api.domain.OrganizationUnits
-import org.jetbrains.exposed.sql.and
-import org.jetbrains.exposed.sql.selectAll
+import io.octatec.horext.api.repository.StudyPlanRepository
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
 @Transactional
-class StudyPlanServiceImpl() : StudyPlanService {
+class StudyPlanServiceImpl(private val studyPlanRepository: StudyPlanRepository) : StudyPlanService {
 
     override fun getAllStudyPlan(): List<StudyPlan> {
-        return StudyPlans.selectAll().map { row -> StudyPlans.createEntity(row) }
+        return studyPlanRepository.getAllStudyPlan()
     }
 
     override fun getStudyPlanById(id: Long): List<StudyPlan> {
-        val sp = StudyPlans
-        val ou = OrganizationUnits
-        return sp.leftJoin(ou)
-        .select( sp.columns + ou.columns)
-        .where { (sp.id eq id) }.map { row -> sp.createEntity(row) }
+        return studyPlanRepository.getStudyPlanById(id)
     }
 }
